move auth + proxy config to transport package
